Detect V1 ptables at startup and log a helpful error. Tests used to use V1 tables for exercising Hash collision paths but now we do this by specifying different hashers

+++ src/EventStore.Core/Bus/InMemoryBus.cs
@@ -80,11 +80,11 @@ public partial class InMemoryBus : ISubscriber, IAsyncHandle<Message> {
 		var elapsedMs = ts.ElapsedMilliseconds;
 		if (elapsedMs > _slowMsgThresholdMs) {
 			Log.Debug("SLOW BUS MSG [{bus}]: {message} - {elapsed}ms.",
-				Name, message.GetType().Name, elapsedMs);
+				Name, message.GetType().Name, (int)elapsedMs);
 			if (elapsedMs > QueuedHandlerThreadPool.VerySlowMsgThreshold.TotalMilliseconds &&
 			    message is not SystemMessage.SystemInit)
 				Log.Error("---!!! VERY SLOW BUS MSG [{bus}]: {message} - {elapsed}ms.",
-					Name, message.GetType().Name, elapsedMs);
+					Name, message.GetType().Name, (int)elapsedMs);
 		}
 	}
